+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Genium project Release Notes
 
 ## Unreleased
+- Breaking changes:
+  + Removed support for Struct ihneritance.
 - Features:
   + Added "Immutable" property support for C++.
 - Bug fixes: